+++ item/components.py
@@ -1,11 +1,3 @@
-from peewee import SqliteDatabase, Model, IntegerField, ForeignKeyField
-from .models import Item
-from settings import DATABASE_NAME
-
-
-database = SqliteDatabase(DATABASE_NAME)
-
-
 class Inventory(object):
     """
     The inventory object containing Items via InventorySlots
@@ -54,20 +46,3 @@ class Inventory(object):
             loaded_inventory.add_item(item)
 
 
-class InventorySlot(Model):
-    """
-    Describes a slot containing an item and referring to an inventory.
-    """
-
-    class Meta(object):
-        database = database
-
-    inventory_id = IntegerField(index=True)
-    item_id = ForeignKeyField(Item, null=True)
-
-    def __init__(self, inventory_id=None, item_id=None):
-        super().__init__()
-        if inventory_id is None:
-            self.inventory_id = InventorySlot.raw("SELECT COUNT(inventory_id) FROM Inventory").execute()
-        else:
-            self.inventory_id = inventory_id
